move some functions to arm-handler-common
